+++ analysis/import/GitLogParser/src/main/kotlin/de/maibornwolff/codecharta/importer/gitlogparser/GitLogParser.kt
@@ -7,6 +7,7 @@ import de.maibornwolff.codecharta.importer.gitlogparser.input.metrics.MetricsFac
 import de.maibornwolff.codecharta.importer.gitlogparser.parser.LogParserStrategy
 import de.maibornwolff.codecharta.importer.gitlogparser.parser.git.GitLogNumstatRawParserStrategy
 import de.maibornwolff.codecharta.model.Project
+import de.maibornwolff.codecharta.serialization.OutputFileHandler
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import de.maibornwolff.codecharta.serialization.ProjectSerializer
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
@@ -19,7 +20,6 @@ import picocli.CommandLine
 import java.io.File
 import java.io.IOException
 import java.io.InputStream
-import java.io.OutputStreamWriter
 import java.io.PrintStream
 import java.nio.charset.Charset
 import java.nio.file.Files
@@ -51,10 +51,14 @@ class GitLogParser(
     )
     private var nameFile: File? = null
 
-    @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File (or empty for stdout)"])
-    private var outputFile = ""
+    @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File"])
+    private var outputFile: String? = null
 
-    @CommandLine.Option(names = ["-nc", "--not-compressed"], description = ["save uncompressed output File"], arity = "0")
+    @CommandLine.Option(
+        names = ["-nc", "--not-compressed"],
+        description = ["save uncompressed output File"],
+        arity = "0"
+    )
     private var compress = true
 
     @CommandLine.Option(names = ["--silent"], description = ["suppress command line output during process"])
@@ -109,13 +113,11 @@ class GitLogParser(
         if (pipedProject != null) {
             project = MergeFilter.mergePipedWithCurrentProject(pipedProject, project)
         }
-        if (outputFile.isNotEmpty()) {
-            if (compress) ProjectSerializer.serializeAsCompressedFile(
-                project,
-                outputFile
-            ) else ProjectSerializer.serializeProjectAndWriteToFile(project, outputFile)
+        val filePath = outputFile ?: "notSpecified"
+        if (compress && filePath != "notSpecified") {
+            ProjectSerializer.serializeAsCompressedFile(project, filePath)
         } else {
-            ProjectSerializer.serializeProject(project, OutputStreamWriter(output))
+            ProjectSerializer.serializeProject(project, OutputFileHandler.writer(outputFile ?: "", output))
         }
 
         return null
@@ -190,7 +192,6 @@ class GitLogParser(
     }
 
     companion object {
-
         private fun guessEncoding(pathToLog: File): String? {
             val inputStream = pathToLog.inputStream()
             val buffer = ByteArray(4096)
@@ -204,6 +205,11 @@ class GitLogParser(
             detector.dataEnd()
 
             return detector.detectedCharset
+        }
+
+        @JvmStatic
+        fun main(args: Array<String>) {
+            CommandLine.call(GitLogParser(), System.out, *args)
         }
     }