+++ analysis/import/SonarImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/sonar/SonarImporterMain.kt
@@ -4,17 +4,14 @@ import de.maibornwolff.codecharta.filter.mergefilter.MergeFilter
 import de.maibornwolff.codecharta.importer.sonar.dataaccess.SonarMeasuresAPIDatasource
 import de.maibornwolff.codecharta.importer.sonar.dataaccess.SonarMetricsAPIDatasource
 import de.maibornwolff.codecharta.importer.sonar.dataaccess.SonarVersionAPIDatasource
+import de.maibornwolff.codecharta.serialization.OutputFileHandler
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import de.maibornwolff.codecharta.serialization.ProjectSerializer
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
 import picocli.CommandLine
-import java.io.BufferedWriter
-import java.io.FileWriter
 import java.io.InputStream
-import java.io.OutputStreamWriter
 import java.io.PrintStream
-import java.io.Writer
 import java.net.URL
 import java.util.concurrent.Callable
 
@@ -28,13 +25,15 @@ class SonarImporterMain(
     private val output: PrintStream = System.out
 ) : Callable<Void>, InteractiveParser {
 
-    @CommandLine.Option(names = ["-h", "--help"], usageHelp = true, description = [
-        "Please locate:\n" +
-        "-    sonar.host.url=https://sonar.foo\n" +
-        "-    sonar.login=c123d456\n" +
-        "-    sonar.projectKey=de.foo:bar\n" +
-        "That you use to upload your code to sonar.\n" +
-        "Then execute [sonarimport https://sonar.foo de.foo:bar -u c123d456]"])
+    @CommandLine.Option(
+        names = ["-h", "--help"], usageHelp = true, description = [
+            "Please locate:\n" +
+                    "-    sonar.host.url=https://sonar.foo\n" +
+                    "-    sonar.login=c123d456\n" +
+                    "-    sonar.projectKey=de.foo:bar\n" +
+                    "That you use to upload your code to sonar.\n" +
+                    "Then execute [sonarimport https://sonar.foo de.foo:bar -u c123d456]"]
+    )
     private var help = false
 
     @CommandLine.Parameters(index = "0", paramLabel = "URL", description = ["url of sonarqube server"])
@@ -48,8 +47,8 @@ class SonarImporterMain(
     )
     private var projectId = ""
 
-    @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File (or empty for stdout)"])
-    private var outputFile = ""
+    @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File"])
+    private var outputFile: String? = null
 
     @CommandLine.Option(names = ["-m", "--metrics"], description = ["comma-separated list of metrics to import"])
     private var metrics = mutableListOf<String>()
@@ -57,19 +56,15 @@ class SonarImporterMain(
     @CommandLine.Option(names = ["-u", "--user"], description = ["user token for connecting to remote sonar instance"])
     private var user = ""
 
-    @CommandLine.Option(names = ["-nc", "--not-compressed"], description = ["save uncompressed output File"], arity = "0")
+    @CommandLine.Option(
+        names = ["-nc", "--not-compressed"],
+        description = ["save uncompressed output File"],
+        arity = "0"
+    )
     private var compress = true
 
     @CommandLine.Option(names = ["--merge-modules"], description = ["merges modules in multi-module projects"])
     private var usePath = false
-
-    private fun writer(): Writer {
-        return if (outputFile.isEmpty()) {
-            OutputStreamWriter(output)
-        } else {
-            BufferedWriter(FileWriter(outputFile))
-        }
-    }
 
     private fun createMeasuresAPIImporter(): SonarMeasuresAPIImporter {
         if (url.endsWith("/")) url = url.substring(0, url.length - 1)
@@ -92,8 +87,12 @@ class SonarImporterMain(
         if (pipedProject != null) {
             project = MergeFilter.mergePipedWithCurrentProject(pipedProject, project)
         }
-
-        if (compress) ProjectSerializer.serializeAsCompressedFile(project, outputFile) else ProjectSerializer.serializeProject(project, writer())
+        val filePath = outputFile ?: "notSpecified"
+        if (compress && filePath != "notSpecified") {
+            ProjectSerializer.serializeAsCompressedFile(project, filePath)
+        } else {
+            ProjectSerializer.serializeProject(project, OutputFileHandler.writer(outputFile ?: "", output))
+        }
 
         return null
     }