+++ analysis/import/CSVImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/csv/CSVImporter.kt
@@ -1,16 +1,14 @@
 package de.maibornwolff.codecharta.importer.csv
 
+import de.maibornwolff.codecharta.serialization.OutputFileHandler
 import de.maibornwolff.codecharta.serialization.ProjectSerializer
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
 import picocli.CommandLine
-import java.io.BufferedWriter
 import java.io.File
-import java.io.FileWriter
 import java.io.IOException
 import java.io.InputStream
-import java.io.OutputStreamWriter
-import java.io.Writer
+import java.io.PrintStream
 import java.util.concurrent.Callable
 
 @CommandLine.Command(
@@ -18,7 +16,9 @@ import java.util.concurrent.Callable
     description = ["generates cc.json from csv with header"],
     footer = ["Copyright(c) 2022, MaibornWolff GmbH"]
 )
-class CSVImporter : Callable<Void>, InteractiveParser {
+class CSVImporter(
+    private val output: PrintStream = System.out
+) : Callable<Void>, InteractiveParser {
 
     @CommandLine.Option(names = ["-h", "--help"], usageHelp = true, description = ["displays this help and exits"])
     private var help = false
@@ -32,8 +32,8 @@ class CSVImporter : Callable<Void>, InteractiveParser {
     @CommandLine.Option(names = ["-nc", "--not-compressed"], description = ["save uncompressed output File"])
     private var compress = true
 
-    @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File (or empty for stdout)"])
-    private var outputFile: File? = null
+    @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File"])
+    private var outputFile: String? = null
 
     @CommandLine.Parameters(arity = "1..*", paramLabel = "FILE", description = ["sourcemonitor csv files"])
     private var files: List<File> = mutableListOf()
@@ -44,20 +44,14 @@ class CSVImporter : Callable<Void>, InteractiveParser {
         val csvProjectBuilder = CSVProjectBuilder(pathSeparator, csvDelimiter)
         files.map { it.inputStream() }.forEach<InputStream> { csvProjectBuilder.parseCSVStream(it) }
         val project = csvProjectBuilder.build()
-        val filePath = outputFile?.absolutePath ?: "notSpecified"
+        val filePath = outputFile ?: "notSpecified"
 
-        if (compress && filePath != "notSpecified") ProjectSerializer.serializeAsCompressedFile(project, filePath)
-        else ProjectSerializer.serializeProject(project, writer())
-
-        return null
-    }
-
-    private fun writer(): Writer {
-        return if (outputFile == null) {
-            OutputStreamWriter(System.out)
+        if (compress && filePath != "notSpecified") {
+            ProjectSerializer.serializeAsCompressedFile(project, filePath)
         } else {
-            BufferedWriter(FileWriter(outputFile))
+            ProjectSerializer.serializeProject(project, OutputFileHandler.writer(outputFile ?: "", output))
         }
+        return null
     }
 
     companion object {